ecotech/ecotech_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
google-generativeai
secure-smtplib